Moved standard libraries to the MetX.Standard namespace
