--- MetX/MetX.Generators.Samples.Client/EmptyClass.cs
@@ -1,5 +1,0 @@
-﻿namespace MetX.Generators.Samples.Client
-{
-    [MetX.Aspects.GenerateAddingStaticCode]
-    public class EmptyClass { }
-}
