+++ MetX/MetX.QuickScripts/QuickScriptEditor.cs
@@ -1,6 +1,7 @@
 ﻿using System.Reflection;
 using System.Reflection.Emit;
 using MetX.Standard.Pipelines;
+using MetX.Windows.Library;
 
 namespace XLG.QuickScripts
 {
