+++ MetX/MetX.Standard.Generators.Samples.Client/EmptyClass.cs
@@ -1,0 +1,7 @@
+﻿using MetX.Standard.Aspects;
+
+namespace MetX.Standard.Generators.Samples.Client
+{
+    [GenerateAddingStaticCode]
+    public class EmptyClass { }
+}
