+++ MetX/MetX.Standard.Generators.Samples.Client/Fred.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace MetX.Generators.Samples.Client
+namespace MetX.Standard.Generators.Samples.Client
 {
     [Aspects.GenerateFromTemplate("exampleOne.template")]
     public class Fred
