+++ MetX/MetX.Standard.Generators/AddClassSupportFromTemplate.cs
