+++ MetX/MetX.Standard.Generators/AddStaticCode.cs
