+++ MetX/MetX.Standard.Generators/GenGen/Operation.cs
