+++ MetX/MetX.Standard.Generators/MetX.Standard.Generators.csproj
@@ -33,7 +33,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\MetX.Aspects\MetX.Aspects.csproj" />
+    <ProjectReference Include="..\MetX.Standard.Aspects\MetX.Standard.Aspects.csproj" />
     <ProjectReference Include="..\MetX.Standard\MetX.Standard.csproj" />
   </ItemGroup>
