+++ MetX/MetX.Standard.Generators/Templates/CSharp/Default/Namespace.AspectsName/GenerateFromTemplate.cs
