+++ MetX/MetX.Standard.Generators/Templates/CSharp/Default/Namespace.AspectsName/GenerateGeneratorName.cs
