+++ MetX/MetX.Standard.Generators/Templates/CSharp/Default/Namespace.AspectsName/Namespace.AspectsName.csproj
