+++ MetX/MetX.Standard.Generators/Templates/CSharp/Default/Namespace.ClientName/Namespace.ClientName.csproj
