+++ MetX/MetX.Standard/Generation/CSharp/Project/ClientCsProjGenerator.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Xml;
-using MetX.Aspects;
 using MetX.Standard.Library;
 
 namespace MetX.Standard.Generation.CSharp.Project
