+++ MetX/MetX.Standard/Generation/CSharp/Project/IGenerateACsProj.cs
@@ -1,5 +1,4 @@
 ﻿using System.Xml;
-using MetX.Aspects;
 
 namespace MetX.Standard.Generation.CSharp.Project
 {
