+++ MetX/MetX.Tests/Standard/Generation/CSharp/Project/Pieces/LangVersion/Missing.xml
@@ -13,7 +13,7 @@
   <!-- From https://github.com/dotnet/roslyn/issues/44093 -->
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <StartupObject>MetX.Generators.Samples.Client.Program</StartupObject>
+    <StartupObject>MetX.Standard.Generators.Samples.Client.Program</StartupObject>
   </PropertyGroup>
   <Target Name="AddSourceGeneratedFiles" AfterTargets="CoreCompile">
     <ItemGroup>
@@ -35,6 +35,6 @@
 
   <ItemGroup>
     <ProjectReference Include="..\MetX.Aspects\MetX.Aspects.csproj" />
-    <ProjectReference Include="..\MetX.Generators\MetX.Generators.csproj" OutputItemType="Analyzer" ReferenceOutputAssembly="false" />
+    <ProjectReference Include="..\MetX.Standard.Generators\MetX.Standard.Generators.csproj" OutputItemType="Analyzer" ReferenceOutputAssembly="false" />
   </ItemGroup>
 </Project>
