+++ MetX/MetX.Tests/Standard/Generation/CSharp/Project/TestHelpers.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 using System.IO;
-using MetX.Aspects;
+using MetX.Standard.Generation;
 using MetX.Standard.Generation.CSharp.Project;
 
 namespace MetX.Tests.Standard.Generation.CSharp.Project
@@ -12,7 +12,7 @@ namespace MetX.Tests.Standard.Generation.CSharp.Project
             var options = CsProjGeneratorOptions
                 .Defaults()
                 .WithFramework(genFramework)
-                .WithPathToTemplatesFolder(@"..\..\..\..\MetX.Generators\Templates")
+                .WithPathToTemplatesFolder(@"..\..\..\..\MetX.Standard.Generators\Templates")
                 .WithOutputPath(Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "Product"))
                 ;
