+++ MetX/deepCleaner.bat
@@ -1,0 +1,24 @@
+@ECHO OFF
+rmdir /q /s I:\OneDrive\Data\code\xlgGallery\xlg\MetX\MetX.Controls\bin
+rmdir /q /s I:\OneDrive\Data\code\xlgGallery\xlg\MetX\MetX.Glove.Console\bin
+rmdir /q /s I:\OneDrive\Data\code\xlgGallery\xlg\MetX\MetX.QuickScripts\bin
+rmdir /q /s I:\OneDrive\Data\code\xlgGallery\xlg\MetX\MetX.Standard\bin
+rmdir /q /s I:\OneDrive\Data\code\xlgGallery\xlg\MetX\MetX.Standard.Aspects\bin
+rmdir /q /s I:\OneDrive\Data\code\xlgGallery\xlg\MetX\MetX.Standard.Generators\bin
+rmdir /q /s I:\OneDrive\Data\code\xlgGallery\xlg\MetX\MetX.Standard.Generators.Samples.Client\bin
+rmdir /q /s I:\OneDrive\Data\code\xlgGallery\xlg\MetX\MetX.Techniques.Tests\bin
+rmdir /q /s I:\OneDrive\Data\code\xlgGallery\xlg\MetX\MetX.Tests\bin
+rmdir /q /s I:\OneDrive\Data\code\xlgGallery\xlg\MetX\MetX.Windows\bin
+
+rmdir /q /s I:\OneDrive\Data\code\xlgGallery\xlg\MetX\MetX.Aspects\obj
+rmdir /q /s I:\OneDrive\Data\code\xlgGallery\xlg\MetX\MetX.Controls\obj
+rmdir /q /s I:\OneDrive\Data\code\xlgGallery\xlg\MetX\MetX.Generators\obj
+rmdir /q /s I:\OneDrive\Data\code\xlgGallery\xlg\MetX\MetX.Glove.Console\obj
+rmdir /q /s I:\OneDrive\Data\code\xlgGallery\xlg\MetX\MetX.QuickScripts\obj
+rmdir /q /s I:\OneDrive\Data\code\xlgGallery\xlg\MetX\MetX.Standard\obj
+rmdir /q /s I:\OneDrive\Data\code\xlgGallery\xlg\MetX\MetX.Standard.Aspects\obj
+rmdir /q /s I:\OneDrive\Data\code\xlgGallery\xlg\MetX\MetX.Standard.Generators\obj
+rmdir /q /s I:\OneDrive\Data\code\xlgGallery\xlg\MetX\MetX.Standard.Generators.Samples.Client\obj
+rmdir /q /s I:\OneDrive\Data\code\xlgGallery\xlg\MetX\MetX.Techniques.Tests\obj
+rmdir /q /s I:\OneDrive\Data\code\xlgGallery\xlg\MetX\MetX.Tests\obj
+rmdir /q /s I:\OneDrive\Data\code\xlgGallery\xlg\MetX\MetX.Windows\obj
